core: introduce DeleteAllVmCheckpointsCommand
The new DeleteAllVmCheckpointsCommand will be used to initiate and
monitor the storage jobs to remove all the VM checkpoints. The command
will ends when all the jobs to remove the checkpoints from the volume ends.

This command will be used for removing all the VM checkpoints after
restoring a VM snapshot due to a possibility for backup corruption.

Change-Id: Id688296781532e3b29d819ee6476766e96329976
Bug-Url: https://bugzilla.redhat.com/1952577
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>