core: disable 'vm_backups' DB table cleanup after backup is over
Disable the automatic cleanup of the backup entry in the
'vm_backups' DB table at the end of the backup command.
This will allow users to see the status of the backup
operation (either via REST API or DB) for a certain
period of time until the scheduled cleanup thread will run.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1849861
Change-Id: I3dee88cc11a6c8f5274a0ca25a1e39d081fb381c