core: Don't show next run memory configuration when unnecessary
When we hot plug memory, we set next run configuration based on the
assumption that the hot plugged memory device will be replaced by base
RAM on next run.  But from the user's point of view it really doesn't
matter, the amount of RAM will be the same on the next run and the
next run configuration displayed for the VM is disturbing or
confusing.

Next run configuration is necessary only when one of the following
conditions applies:

- The user asks to apply the change later, not immediately.

- The amount of memory is reduced and it cannot be achieved with
  memory hot unplug, either because memory hot unplug is not supported
  or the requested amount is lower than the amount of memory when the
  VM was started.

We don't have access to the VM's original memory to check for the very
last condition above.  But it actually doesn't matter that much
because the Web UI doesn't permit memory hot unplug by editing the
amount of memory, it's only possible to unplug a memory device.  So we
can signal next run configuration whenever the amount of memory is
reduced.  The only problem arises when memory is hot unplugged via
REST API, next configuration is signaled in such a case.

Alternatively, we could remove the memory check for next run
configuration completely.  But in such a case, next run configuration
is not signaled even when Apply Later is selected.

In any case, the updated check assumes all the valid memory hot plugs
succeed.  If the corresponding hot plug request fails then the amount
of memory should not be updated at all.  If the request succeeds but
the action fails in the guest OS, we cannot fully detect and handle
such a case and assume success as for the next run configuration.

Change-Id: I1d0208c0165d9db7668939b66aa9f9377e8abc8c
Bug-Url: https://bugzilla.redhat.com/1779983
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>