core: add phase validation for the 'backup stop' command
Backup can be stopped and finalized only in 'READY' state.
Adding validation that the backup is not being stopped
while it is either:
1) WIP, i.e., during 'CREATING_SCRATCH_DISKS' or 'STARTING' phases.
2) When already finished (either successfully or with failure).

Error message will look like:
Cannot stop VM backup. The VM backup is not in READY phase, backup phase is CREATING_SCRATCH_DISKS. Please try again when the backup is in READY phase.
Cannot stop VM backup. The VM backup is not in READY phase, backup phase is SUCCEEDED. Please try again when the backup is in READY phase.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Change-Id: I0e0d013f0a00ce5f96dff98fd6da31faf9c97ba9