core: update engine validations in backup flows
- Updating engine validations due to adding 2 new
finishing phase statuses and canceling database cleanup
after the backup was finalized.
Now, in order to check that there is no backup currently in
progress, it's not enough to check that there are no backup
entries in the database, but need to take into account also
the statuses of the existing entries. I.e., "SUCCEEDED" and
"FAILED" statuses mean that the backup has already finished.

- Using the same error message to indicate 'VM is during backup'
instead of the previous situation where 2 almost identical
messages were used across the code and the name of one of them
was confusing: i.e., using "CANNOT_START_BACKUP_ALREADY_IN_PROGRESS"
inside the "DeleteVmCheckpointCommand" implementation).

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1849861
Change-Id: I0b134d84a081e1e9068f8f69dbf1a20d52b1ec02