webadmin: expose logging to plugins
This patch allows the plugins to use the logging
mechanism configured for the gwt webadmin. It allows
to have unified approach for both, the webadmin and plugins.

This patch also adds the possibility to handle uncaught exceptions
in the same way as the webadmin does - the message is remote
logged into ui.log file and a vojtobox is displayed.

Change-Id: Ifb93a601025ffd20c4458310436537ad605267df
Bug-Url: https://bugzilla.redhat.com/1771925
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>