webadmin: calculation of max memory in VM dialog
When the max memory is calculated in the VM dialog,
the calculated value can be greater than the max memory
size defined in the vdc_options table. Then the validation
of the dialog fails.

This patch calculates the memory in the same way as before,
but if the calculated value is greater than the vdc_options one,
the max memory in the VM dialog is set to the value from the
vdc_options.

Change-Id: I3351b06e4a80b6d9dfba47879d278d881c79ba76
Bug-Url: https://bugzilla.redhat.com/1946876
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>