core: Don't always power off on shutdown during migration
When a VM is migrating and its shutdown is requested, Engine calls
power off unconditionally.  This is because clean shutdown is allowed
only when the VM is in Up status.  When the VM is not Up we don't know
whether it can react to a normal shutdown.

This is unnecessarily harsh.  Let's soften the condition and allow
clean shutdown also when the VM is migrating and it was in Up status
before it started to migrate.  We track the last status before
migration in VmManager because it's accessible to both migration and
shutdown commands and tracking the status in the database is probably
not worth it.

Another alternative might be to always call a clean shutdown on
migrating VMs, with force=True.  Then if the shutdown is unsuccessful
within the given timeout, the VM will be powered off.  This might be
simpler and more universal but it could, at least in theory, cause
trouble for VMs with slow shutdowns exceeding the timeout.

Change-Id: I8bb51c3a87918b7ce097ed400adcaab4f37fc92c
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>