core: use disk's sd_id when retrieving info
If a template is imported with disks residing on different SDs, it will
try to query the size of the disk on the SD on which the import
operation is run, regardless of where the disk actually resides.

This patch uses the disk's Storage Domain ID when querying vdsm about
its size.

Bug-Url: https://bugzilla.redhat.com/1970718
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I8116f2094f2646f0baa656ffefc656d19db4316f