core: Validate OS compatibility with Q35 on VM import
When a VM with Q35 chipset is imported, validate that its guest OS
supports Q35, and fail the import if it does not.

When a VM with I440FX chipset is imported into Q35 cluster, do not
change the chipset of the VM to Q35, if its guest OS does not support
Q35.

Change-Id: Ib3759acf9b3d50a87b0efc32cf0167244e94bea2
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>