restapi: Setting bochs display with REST API
REST API does not allow currently to set display type separately from
graphics type of a VM. The display type is chosen automatically by the
graphics type using the list of supported display-graphics combinations
defined in osinfo. Before introduction of bochs display type, VGA was
preferred for VNC graphics type.

This patch adds a special case when bochs display type is automatically
chosen. For this to happen, the following conditions are to be met:

1. The VM should be configured with UEFI BIOS type.
2. The VM should be of 'Server' type.
3. The VM should use VNC.
4. VNC/bochs combination should be present in osinfo for the
   corresponding cluster compatibility version.

Change-Id: If0dd41e75c5be349fd93e442fd44c0b8cf4325a6
Bug-Url: https://bugzilla.redhat.com/1947312
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>