Fix the ovirt-host-deploy role, kernel options issue
The code was previously updating only the default kernel (boot loader entry that is configured to boot by default), thus the changes were not surving kernel updates,
 even on plain RHEL/CentOS/CentOS Stream, as doing that does not update the default kernel options of the bootloader.

The patch modifies the role, so that the kernel options will be added to the default conf. as well, and thus will be added to newly installed kernels on RHEL/CentOS/CentOS Stream.
It will also modify the configuration for all currently installed kernel versions.

Change-Id: Ice0b71fdf7ce53928cc45ce8446c70cbcdff2f56
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1972023