webadmin: remove foreman hosts provisioning
This patch removes the ability to discover / select already provisioned
hosts (satellite) and register them as ovirt hosts.

Please note  that errata handling is expected to be left intact. That
means that host must be registered and provisioned on the satellite side
and FQDN must match on both sides. Ovirt engine must only store
information about the provider id (with VDS data in postgres) so that it
knows how to fetch erratas via HTTP call (using Katello/Foreman API).

Bug-Url: https://bugzilla.redhat.com/1901011
Signed-off-by: Artur Socha <asocha@redhat.com>
Change-Id: Ie7d1e44baff030f4940f8ea906b62f1ad5500981