core: Don't pretend an unaligned memory hot plug is applicable
Memory hot plug must be aligned to the architecture specific minimum
DIMM size.  If it is not then the hot plug fails.

We already have a check for this in UpdateVm command validation.
However it is already a bit too late when using the web UI.  When a
user increases Memory Size in the VM edit dialog and asks to apply the
change immediately rather than on the next run then the hot plug will,
get rejected, Memory Size will be reverted to its original value and
Physical Memory Guaranteed will keep the new updated value.  The user
is confused and Memory Size and Physical Memory Guaranteed won't
mutually match.

Let's check for the misalignment already in the update confirmation
dialog and let's not offer memory hot plug when the memory increase is
not aligned.  Let's offer only next run change in such a case.

When using REST API to attempt a misaligned memory hot plug, Memory
Size is updated, next configuration is not set, no error is reported
and the hot plug is not performed.  Which doesn't look entirely right
but a contingent fix is for another patch.

Change-Id: I5b34a2b6b926222779a5b2635c1c8de547782e88
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>