webadmin: Improve "TPM device is not supported" message
When a user attempts to enable a TPM device in the VM edit dialog for
a VM not having the required compatibility level, he gets an error
message with a list of reasons.  The problem with the compatibility
level is not listed there, which is quite confusing.  Let's add it.

Let's also add a missing ${type}.

Change-Id: I7fd40f03d02d4751be657507258ed8db7052560d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>