core: disable 'image_transfers' DB table cleanup after image transfer is over
Disable the automatic cleanup of the image transfer entry in the
'image_transfers' DB table at the end of the backup download command.
This will allow users to see the status of the image transfer
operation (either via REST API or DB) for a certain
period of time until the scheduled cleanup thread will run.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1849861
Change-Id: I4749d531fd7568cf848f07ad03a9e52f511a22a3