core: free memory locks before ending job
Currently, the memory locks are released after the job is marked as
finished in the database. However, many clients of ovirt's API rely on
the job's status to determine whether an operation is complete (namely
OST).
Releasing the memory lock after marking the job is finished leaves a
small window of opportunity for race conditions to occur, where the job
is seen as finished by a client yet the lock is still held.

This patch reorders the operations to free the memory lock before
marking the job as finished to prevent these type of races.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: Iac8d69a426ce0d7958ec732c158bc1aa80200c56