core: Improve memory hot plug/unplug size checks
Hot plugged memory size must be dividable by an architecture specific
value.  This is currently not strictly checked in UpdateVmCommand
validation, the hot plug is only prevented from running if the size
alignment is not satisfied.  It is possible to hot plug an unaligned
amount of memory using REST API without setting next run
configuration, resulting in the hot plug not being performed, memory
amount for next run configuration changed and next run configuration
not set, which is wrong.  There are also other validation and next
configuration problems.

The patch makes the following memory hot plug and unplug improvements:

- The amount of hot plugged memory is properly validated now.

- Memory hot plugs and unplugs are prevented when memory next run
  configuration already exists, to avoid confusions about amounts of
  hot plugged or unplugged memory.

- The memory next run configuration check looks at the hot plugged
  size now (this shouldn't be needed with the validation improvement
  but let's have the check complete).

- HotPlugMemoryBlockSizeMb config value is removed because it's no
  longer used and is not needed when we already have architecture
  specific checks.

We still allow attempts of memory hot unplugs below the VM starting
memory, because there is no way to get the starting memory.  A proper
validation should look at the available DIMMs, which is a different
topic.

Change-Id: Ibf9783424424baeebfe7d039262e3e64b46e71d4
Bug-Url: https://bugzilla.redhat.com/1982065
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>