webadmin: fix resume image upload from UI
Following the fix for starting an upload from the UI not working,
the resume operation was affected, which caused the operation to
run in the wrong way and get stuck on the confirmation window.

The fix should apply to new uploads only and not on the resuming operation.

ScreenShots: https://imgur.com/a/drsD2Vj

Change-Id: Ia953299bfaffd575328e757f833b5f009d6db710
Bug-Url: https://bugzilla.redhat.com/1950593
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>