webadmin: fix runtime exception for LUNs wrong casting
From the UI, when displaying disk view with all the disks
in the system, Luns disks got runtime error when the curser
selected storage domain filed.

Under the function getTooltip(Disk object), all disks are cast
to diskImage type and the field of the storage domain is used
for displaying, this causing an exception for the wrong casting
of LUN disks.

This patch checks the disk type and using the right field and
casting accordingly.

ScreenShot: https://imgur.com/a/PpR3RyX

Change-Id: I02f08509ccbdd54f0b1d633ac3a1bcdedb59c4a8
Bug-Url: https://bugzilla.redhat.com/1853501
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>