frontend: fix memory stats for OGA
We originally displayed the stats as three values "free/buffers/caches",
but recently we did a change where we used only two values
"free/buffers+caches" if the information comes from QGA and retain the
old format when the information comes from OGA. But this change broke
the UI and now there is nothing displayed when the information comes
from OGA. This patch should fix that.

Since the statistical data isn't provided on import, unnecessary code
removed.

Change-Id: I7ad74a7c0fd713d98d4c0a7f3a0079d4ffe9fb32
Bug-Url: https://bugzilla.redhat.com/1978253
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>