core: add affinity modification commands
This patch will add the commands for editing an affinity group. Until
now we deleted the whole group, which cascaded the members of this
group. Then, it created the same group again, adding the members after
the modification - adding or removing an entity from the affinity group.
Now, we have a command to add or remove only the entity without the
recreation of the affinity group.

It will be used in the API layer, which reported to have a race when
making multiple changes on the affinity group in rapid speed, hitting a
duplicate key violation on the affinity group entity.

Change-Id: I97f7d6139b2703b2c1b02f2d78f9ae24da2e2d6c
Bug-Url: https://bugzilla.redhat.com/1950767
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>