restapi: change the affinity calls
Until now we always edited the whole affinity group entity even when
modifying only the members in it. This patch will call a specific
command for each action, leading to specific change in the DB in order
to prevent a race when having multiple calls on the whole affinity
group.

Change-Id: Ie86d89236d1a91c52b7d0f642c1f8be68dfcc051
Bug-Url: https://bugzilla.redhat.com/1950767
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>