core: pack_ova: wait for loop device
The `Resource temporarily unavailable` error we hit from time to time
when exporting an OVA is a known issue. There is a fix in kernel-utils
https://www.spinics.net/lists/kernel/msg3977449.html .
We should try to solve it on our end for users who didn't get that fix
yet. In this patch we will try to wait 10 seconds to get the loop device
available.

Change-Id: I2a5812cf9a640909ae8553ad9d63a5f5592ca9c4
Bug-Url: https://bugzilla.redhat.com/1975225
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>