engine: Make VM sealing configurable
Make possible to control VM sealing with 'seal' parameter, when creating
a VM, a VM pool or adding new VMs to a pool.

In WebAdmin, added "Sealed" checkbox in "New VM", "New Pool" and "Edit
Pool" dialogs. REST API will still use the default behaviour, defined in
the commit 950952f5d5213b8248f233b8f89304d7b7c05da3.

The "Seal" checkbox is checked by default, if a VM or a stateful VM Pool
is created from a sealed template. It is unchecked by default, if a
stateless VM Pool is created. It is unchecked and disabled, if Windows
is selected as a guest OS.

Change-Id: I6dc14d30aadc15a32a0ccd24f1a4311611f45908
Bug-Url: https://bugzilla.redhat.com/1973270
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>