webadmin: apply changes later when no hot-plug is set
VmNextRunConfigurationModel defaults to 'apply changes immediately' and
the UI enables the user to override if the changes should apply later
(on reboot). But if the hot-plugged panel is hidden, the user is told
that there's no device/setting to hot-plug/set and doesn't have the
option to apply the changes later so the default of 'apply changes
immediately' remains. But when we look at it from the backend point of
view, it's like the webadmin asks the engine to apply the specified
changes now - and that's problematic if those changes cannot be applied
while the VM is running.

Change-Id: I379fccd5cbd19ba22a53015474bf8e6168bdb595
Signed-off-by: Arik Hadas <ahadas@redhat.com>