core: fix the 'images_storage_domain_view' to show disks only once
After the change that was performed lately in "image_transfers"
DB table (the transfer entities are not removed immediately after
the command finishes but after 15-30 minutes), this caused multiple
issues related to returning the same disk multiple times.
Some of the issues were UI-presentation only, some really caused
operations to be performed multiple times on the same disk as-if
indeed there were many different disks.
This patch fixes it - only disks having a not final
"image_transfers.phase" are returned: 9 is "FINISHED_SUCCESS",
10 is "FINISHED_FAILURE".

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1980428
Change-Id: I52d0ec08e9a36fe59e4ad4c3ac57eb0d3be4fb1e