core: Current VCPUs mustn't exceed max VCPUs
The maximum possible number of VCPUs depends on the number of threads
per CPU socket, thus variable.  When the requested number of VCPUs is
close to the possible maximum, it may happen that it exceeds the
variable limit.  Then the combination of the values is invalid in the
domain XML and the VM fails to start.

Let's catch such a situation already in Add/UpdateVmCommand
validation.  The web UI should be further improved, in followup
patches, to assist the user with not exceeding the maximum number of
VCPUs for the given CPU configuration.

Change-Id: I0e054ea66c17dec6427e25d2b78048f72e4f4ea8
Bug-Url: https://bugzilla.redhat.com/1946231
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>