core: Fix missing jackson-databind module dependency
As a part of https://gerrit.ovirt.org/108296 it was added JsonNode class
from jackson-databind module to JsonHelper class. The build went fine
because jackson-databind module was declared as a dependency in top
level maven project, but on runtime there was raised WELD-000119
warning to not generate bean definitions for JsonHelper class because of
missing JsonNode class dependency.
This patch adds jackson-databind module runtime dependency to utils
module.

Change-Id: I1ec2b572eff5c3aea4e9017d473d96fda6b4fa69
Signed-off-by: Martin Perina <mperina@redhat.com>