core: allow specifing VM backup description
Adding a new 'description' field to VM backup entity.
This new field will give the user the option to add a description
up to 1024 characters on the created backup.

This description will allow backup vendors to identify the backup
created by their backup application and add additional info on it.

Change-Id: Iea9c6c083cc59a161dd31b76025d6db02de45f84
Bug-Url: https://bugzilla.redhat.com/1963083
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>