core: allow adding 'description' to VM checkpoint
Adding a new 'description' field to VM checkpoint entity.
This new field will give the user the option to add a description
up to 1024 characters on the created backup.

This description will allow backup vendors to identify the backup
created by their backup application and add additional info on it.

The description will be initialized when the backup starts, the
description of the backup and the checkpoint will be the same.
It should be persisted in the VM backup checkpoint entity
because the VM backup entity was removed from the database eventually.

Change-Id: I566e759c3d95eb9621381a030a14cfdd4f411dee
Bug-Url: https://bugzilla.redhat.com/1963083
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>