backend: add cpu pinning policy to DB
This patch adds the cpu_pinning_policy into the DB. The auto pinning
policy already existed but now it will be saved on the DB.
Currently, `Pin` policy removed. On a following patch the saved value
will be used to set the policy on a scheduled host and it will be
settable also on template entity.

Change-Id: I15f874a16e9a223fe668d371478a219100ac4df2
Bug-Url: https://bugzilla.redhat.com/1979041
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>