core: switch cpu pinning to run phase
Previously we selected a "random" host from the dedicated host list and
used its hardware to configure the cpu pinning policy. In this patch
the configuration phase moved from the AddVmCommand and UpdateVmCommand
to RunVmCommand. We now configure the VM based on the host it is
scheduled to run on.

Now we have the `current` CPU fields. These are set into VmDynamic and
will be determined when the VM starts to run and then will be consumed.
Once the VM is down these fields will reset, while the static
configuration stays.

Change-Id: I7ce0323776a16b307e9b22981fd8bb2f6eed623b
Bug-Url: https://bugzilla.redhat.com/1979041
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>