core: allow set host to maintenance if image transfer is not active
Until now host can be set to maintenance if there was an image
transfer in 'pause' state.

Since bug - https://bugzilla.redhat.com/1981297 fixed, image transfer
entities remain in the system even after the transfer session ended.

This patch allows the host to be set to maintenance even if there are
image transfers on the system that finished successfully/failed.

Change-Id: Icba533c9dc46f9802c148602404e79cdd0bcdd58
Bug-Url: https://bugzilla.redhat.com/1987295
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>