core: add '_update_date' column to 'vm_backups' DB table
Add and use a new date column that will store the date
when the backup was last updated.

The date is needed for the correct cleanup of the backup entry
by a dedicated thread (i.e., 15 minutes for backups that were
finished successfully).

Currently the cleanup is performed according to the "_create_date"
column, which obviously is the wrong date to be used,
and in case of long backups can result in a too early cleanup
(before the user had a chance to retrieve the backup result).

The internal DB '_update_date' column is presented at code and
API/SDK by a more readable 'modificationDate' naming.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1983636
Change-Id: I03a7f70094edf3f63d609b669773880046eb3bcd