core: consider vCPU / physical CPU count when scheduling
Sometimes, it might be useful to consider not only
the CPU utilization of a host, but also the ratio
between the sum of vCPUs of the VMs on the host
to the actual physical CPU cores count.

To allow this, a new parameter was introduced to the
cluster policy called VCPU_TO_PHYSICAL_CPU_RATIO. When
set to 0, the ratio won't be considered (the current state
would be preserved).

When set between 0 and 2.9, if two hosts have the same
CPU utilization, the one with least vCPU/physical CPU
ratio would be preferred.

When checking the load balance, the host with the vCPU/physical CPU
ratio greater than the preconfigured VCPU_TO_PHYSICAL_CPU_RATIO
would be considered as over utilized and some VMs might
be selected to migrate to another host.

Change-Id: If0b07493a982d351e93c3907e9b46b88425c090c
Bug-Url: https://bugzilla.redhat.com/1849169
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>