webadmin: improve the error message when finalizing the finished backup
Until now trying to finalize the backup that was already finished
(with either success or failed status) resulted in a somewhat
awkward message like the following one:

Cannot stop VM backup.
The VM backup is not in READY phase, backup phase is SUCCEEDED.
Please try again when the backup is in READY phase.

The message makes sense when backup is in "INITIALIZING" or
"STARTING" phases, but looks strange when the backup is over or
FINALIZING, because obviously the backup phase will never get back
to the "READY" phase. Removing the confusing last sentence.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Change-Id: I89ac756b58cdb042c12250cb872890957e8de545