core: cleanup in VmAnalyzer
Move the part that handles the data on NICs from the guest agent to
be closer to where VmDynamic is updated in VmAnalyzer. It is safe
since if the hash of this data has changed then 'isAnyFieldChanged'
that also checks this field would always return true.

This patch also cleans up the handling of IP addresses from
Kubevirt in VmAnalyzer.

Change-Id: Ida7632404099a7c63ade6ddfaf1dc73ed2139c3a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1901572
Signed-off-by: Arik Hadas <ahadas@redhat.com>