core: fix handling of guestAgentNicsHash
We must avoid updating the hash of the information on NICs we get from
the guest agent without persisting the data into the database (including
the updated vm_dynamic#vm_ip). Otherwise we risk having them out of
sync, for example:
1. The engine clears vm_ip and guest_agent_nics_hash when destroying a VM
2. The engine "unexpectedly" detects the VM running on another host
3. The engine only updates guest_agent_nics_hash without updating vm_ip
4. vm_ip would not get updated anymore until guest_agent_nics_hash changes

This patch makes sure that when detecting that VM runs "unexpectedly" on
a different host and guest_agent_nics_hash is updated, vm_ip and the
vm_guest_agent_interfaces table would also be updated. It may not be
necessary when detecting the VM in MigratingTo status anymore (since we
should not have "unmanaged" migrations) but better be safe than sorry.

Change-Id: I05224320f9aea78eefdbb284c485e966fc013ff2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1901572
Signed-off-by: Arik Hadas <ahadas@redhat.com>