core: persist guest agent nic info before persisting its hash
We have to persist the data before persisting its hash because otherwise
we risk having them out of sync: if the hash is persisted and the engine
goes down before persisting the data, it would restore the hash when it
boots and unless the data (and therefore its hash) changes, we won't
have the latest data in the database.

We could have updated everything within a transaction to ensure the data
and its hash is always in sync but we want to minimize the use of
transactions by VMs monitoring.

Therefore, this patch moves the part that persists the data to take
place before persisting its hash.

Change-Id: I032429e954d4759722b43ed41a62cce640497897
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1901572
Signed-off-by: Arik Hadas <ahadas@redhat.com>