core: Plug video devices on import
In RHV 4.3, if a VM without a plugged video device is started, the VM
gets an unmanaged cirrus device. But in RHV 4.4 it does not, making
impossible to open the console. If user is unaware about unplugged video
device in the VM, this behaviour may come unexpected after exporting a
VM in RHV 4.3 and importing it in RHV 4.4.

This patch marks all video devices as plugged at VM import.

Change-Id: Ibba26be5d801548368aea892deb4b7f8276e461d
Bug-Url: https://bugzilla.redhat.com/1958398
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>