restapi: on add vm, reset template quota when input quota empty
When adding new vm from template from web-ui, there is no
option to select quota, so it takes always the template quota.

This caused situation that user can't add a VM if not
authorized for the template quota, without having the
option to change it.

This change sets in the rest api the template quota id
to null in case quota sent without id.

then the quota will be set to one that the user have
authorization for(handled by previous patch[1] that
fixed adding from blank template in web ui use case)

[1]: https://gerrit.ovirt.org/c/ovirt-engine/+/94953

Change-Id: I774f46a4e838b4cee87bca95b3c19d66f0f14091
Bug-Url: https://bugzilla.redhat.com/1959186
Signed-off-by: Saif Abu Saleh <sabusale@redhat.com>