restapi: add cpu pinning policy to vmbase
This patch will allow consuming and viewing the value of the cpu
pinning policy via the API on the VM and template entities.
It will prefer the new value on adding new VM. When the VM running the
CPU topology and pinning will be shown under dynamic_cpu.

Since AutoPinningPolicy is deprecated, in case we have input of both CPU
and Auto Pinning Policy, the CPU will be preferred.

Change-Id: Ie133b62996df0f1535579757b2a5ec65922dc584
Bug-Url: https://bugzilla.redhat.com/1979041
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>