core: catch all VmStats exceptions
If PollVmStatsRefresher#poll throws an exception the scheduled thread
will exit and will no longer monitor VMs until engine is restarted. This
can happen if engine loses the connection to the database.

This patch catches all throwable exceptions that may be thrown during
polling, ensuring the thread will be able to recover once the error is
fixed (say, the database connection is restored).

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1770027
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: Ia3bdd38e6a32e74f6f08dbe6480d17728829ab34