spec: Do not require otopi-devtools
bundle-ovirt-engine-health needs otopi-bundle.
When bundle-ovirt-engine-health was introduced, otopi-bundle was part of
otopi-devtools, so otopi-devtools was added to the spec file. Later, it
was moved to a new package otopi-common, as part of the work to support
python2 and python3 in parallel. Recently, for [1], otopi-devtools was
completely removed, as it's not needed anymore, containing only
java-related code which isn't used anymore. So: update the spec to not
require otopi-devtools anymore.

[1] https://bugzilla.redhat.com/1983047

Change-Id: I837b688319f018ecf6290b6ddfda410f088d4459
Signed-off-by: Yedidyah Bar David <didi@redhat.com>