core: avoid passing lock to child command for image transfer
When performing any upload/download action on a disk during the image
transfer command, the memory lock on the disk is released before time.
In the current implementation, the lock is passed to the child command
TransferImageStatusCommand, which has no use of the lock, and releasing
it automatically before the actual transfer is complete.
If the user will try to delete the disk in this period of time
he will succeed and an NPE will be raised since the disk information is
not fully saved yet.

This patch fixes the locking mechanism by not passing the lock to the child
command, TransferImageStatusCommand, by that we prevent the premature release
of the lock before the main command is fully completed.

Change-Id: Ibf3e2305dd0c8bb60c732d5268ff02221d4c9b43
Bug-Url: https://bugzilla.redhat.com/1966535
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>