core: skip snapshot creation for Export VM to SD
When Export VM with "Collapse snapshots" unchecked is triggered,
a snapshot is created on source VM in order to allow for the clone VM
operation to be performed on a snapshot rather than on an active
volume - this allows the source VM to keep running during the clone
operation. Then, all the snapshots are copied to the destination VM,
and then the snapshot created for the operation on source VM is
removed.

The bug was opened for the copy of this snapshot that was
not removed from the destination VM. Since this happens only for
Export VM to Storage Domain No Collapse Snapshots action,
and running VM during this procedure is not allowed anyway - there
is no need to create this snapshot in the first place.

Change-Id: Idf59fe34977fd1e3c96ea807d97007864664286c
Bug-Url: https://bugzilla.redhat.com/1959141
Signed-off-by: Mark Kemel <mkemel@redhat.com>