conf: remove redundant timezone config files
This files was used to validate that the time zone
config values that was set with engine-config is
valid.

The validation check mechanism relied on staticly
configured time zones on the engine side which does
not take into account dynamically added timezones.

After the patch that changed timezones config type
to string[1], this files is no longer needed.

[1]: https://gerrit.ovirt.org/c/ovirt-engine/+/116092

Signed-off-by: Saif Abu Saleh <sabusale@redhat.com>
Change-Id: I2142b1833fa7c7d6d8b405d61dc5d27cc07dca22