core: Increase the maximum number of CPU sockets
The traditional limit on maximum 16 CPU sockets makes no longer
sense.  According to
https://bugzilla.redhat.com/show_bug.cgi?id=1946231#c14, we can drop
the limit.

It doesn't mean that everything will work well with arbitrary number
of CPU sockets, e.g. the guest OS may not support it.  But we've
already been preferring setting a high number of CPU sockets over a
high number of cores or threads, just within the limit of 16.  Users
may override the default N-1-1 sockets-cores-threads setting in the VM
configuration if it causes trouble.

Let's effectively remove the limit on the maximum number of CPU
sockets for 4.6 cluster level.  There are no bugs requesting such a
change for older cluster levels so let's be conservative and don't
touch them to avoid contingent troubles with older QEMU versions.

One side effect of this change is that the maximum number of vCPUs is
no longer limited by the maximum number of CPU sockets.  This doesn't
come without price, setting higher maximum vCPU values takes more
memory from the guest OS, even if those vCPUs are unplugged.  For this
reason, new configuration option MaxNumOfCpusCoefficient is
introduced, limiting the maximum number of vCPUs to about twice the
initial number of vCPUs (considering alignments and the former limits
induced by the maximum number of CPU sockets).

We still retain the limits on the number of CPU cores
(254; https://bugzilla.redhat.com/1505576) and threads
(1/8; https://gerrit.ovirt.org/47394).

Change-Id: I1024442e6a27f7762173ed2ede98d7149c998272
Bug-Url: https://bugzilla.redhat.com/1946231
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>