core: Align disks in OVA to 4K boundary
When exporting a VM to an OVA located in an NFS directory, aligning
disks in the OVA to 4K boundary gives a significant gain in the
performance of writing. Since tar file format does not allow to insert
padding longer than 512 bytes between files, pack_ova.py script creates
zero-filled files of the needed length that are used for padding. On
import, these files are ignored.

Change-Id: I364ae367f813ad2e33b2e024c6a55ceaa24d964c
Bug-Url: https://bugzilla.redhat.com/1927985
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>