core: serialize executions of AttachDiskToVmCommand
We serialized the execution of the code that looks for a free
PCI address for disks in order to avoid a conflict of allocating
the same address to two different disks that are plugged at the
same time (I1292574d3dd32271249d83d1dddb5afabe3672fe).

However, it is not a complete solution because when AttachDiskToVm
is executed, it first loads the disks from the database and then
go over them to find an address that is available. But only the
second part is within the critical section so the following race
may happen:
1. 1st instance of AttachDiskToVm loads the disks of VM V
2. 2nd instance of AttachDiskToVm loads the disks of VM V
3. 1st instance of AttachDiskToVm finds available address and
   allocates it to disk D1
4. 2nd instance of AttachDiskToVm is not aware of the disk D1 that
   was attached and to the allocation made the in previous step
   and therefore allocates the same address to disk D2
5. 2nd instance of AttachDiskToVm fails due to duplicate address

Applying the same solution as in HotPlugDiskToVmCommand where the
disks are loaded from the database within the critical section is
problematic in AttachDiskToVmCommand because the disks from the
database are used in a preliminary validation step of the command.

The solution presented in this patch is to consider the entire
execution of AttachDiskToVmCommand as a critical section instead.

Change-Id: Ia4a89c6d3e595f50e6f1821bfe1665c238b56d7d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1989794
Signed-off-by: Arik Hadas <ahadas@redhat.com>